requirements.txt
attrs==22.2.0
coverage==7.1.0
exceptiongroup==1.1.0
iniconfig==2.0.0
packaging==23.0
pluggy==1.0.0
pytest==7.2.1
pytest-cov==4.0.0
python-mpd2==3.0.5
tomli==2.0.1
tomlkit==0.11.6
